refactor jruby extension in attempt to fix arcball jruby-9.0.0.0
