+++ JRakefile
@@ -25,7 +25,7 @@ if RUBY_PLATFORM =~ /java/
   # -*- ruby -*-
   require 'rake/javaextensiontask'
   
-  Rake::JavaExtensionTask.new('processing') do |ext|
+  Rake::JavaExtensionTask.new('monkstone') do |ext|
     jars = FileList['lib/*.jar']
     warn WARNING unless FileTest.exist? 'lib/video.jar'
     ext.classpath = jars.map {|x| File.expand_path x}.join ':'
