+++ ext/monkstone/fastmath/DeglutLibrary.java
@@ -1,4 +1,4 @@
-package processing.vecmath.vec2;
+package monkstone.fastmath;
 
 import java.io.IOException;
 import org.jruby.Ruby;
@@ -9,7 +9,7 @@ import org.jruby.runtime.load.Library;
  *
  * @author Martin Prout
  */
-public class Vec2Library implements Library{
+public class DeglutLibrary implements Library {
 
     /**
      *
@@ -19,6 +19,8 @@ public class Vec2Library implements Library{
      */
     @Override
     public void load(final Ruby runtime, boolean wrap) throws IOException {
-        Vec2.createVec2(runtime);
-    }  
+        Deglut.createDeglut(runtime);
+    }
+    
 }
+
