+++ library/fastmath/fastmath.rb
@@ -1,4 +1,4 @@
 require_relative '../../lib/ruby-processing'
 require "#{RP5_ROOT}/lib/rpextras"
 
-Java::ProcessingFastmath::DeglutLibrary.new.load(JRuby.runtime, false)
+Java::MonkstoneFastmath::DeglutLibrary.new.load(JRuby.runtime, false)
